listener/requirements.txt
fastapi==0.55.1
uvicorn==0.11.3
pydantic==1.4
psycopg2-binary>=2.7
